Kill off a l o t of old files
