+++ src/App.tsx
@@ -11,8 +11,8 @@ import { colorsV3, fonts, getCdnFontFaces } from '@hedviginsurance/brand'
 import { CssBaseline } from '@material-ui/core'
 import { MuiThemeProvider } from '@material-ui/core/styles'
 import { history } from 'clientEntry'
-import Breadcrumbs from 'components/shared/navigation/breadcrumbs/Breadcrumbs'
-import { VerticalMenu } from 'components/shared/navigation/sidebar/VerticalMenu'
+import { Breadcrumbs } from 'features/navigation/breadcrumbs/Breadcrumbs'
+import { VerticalMenu } from 'features/navigation/sidebar/VerticalMenu'
 import { Routes } from 'pages/routes'
 import React, { useState } from 'react'
 import { hot } from 'react-hot-loader/root'
