+++ src/features/navigation/sidebar/VerticalMenu.tsx
